[feature][test] Enhance utility functions and add comprehensive tests for model training and error handling

+++ src/admet/train/base/utils.py
@@ -15,7 +15,7 @@ logger = logging.getLogger(__name__)
 
 
 def _extract_features(df, fingerprint_cols: Sequence[str]) -> np.ndarray:
-    return df.loc[:, fingerprint_cols].to_numpy(dtype=float)
+    return df.loc[:, fingerprint_cols].to_numpy()
 
 
 def _extract_targets(df, endpoints: Sequence[str]) -> np.ndarray:
